Converting LatLng types to Double; Fix ? bug
A bug was found when trying to encode polylines that contained question marks (?).
This occurs when there is no difference between two subsequent points.

+++ src/main/scala/LatLng.scala
@@ -1,6 +1,6 @@
 package com.trifectalabs.polyline
 
 case class LatLng(
-  lat: BigDecimal,
-  lng: BigDecimal
+  lat: Double,
+  lng: Double
 )
